Move everything into SubscriptionService
